🔄 refactor codes
